# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Swans ShuttleHub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stops:
    get:
      operationId: listStops
      tags: [stops]
      summary: List all stops
      responses:
        "200":
          description: List of stops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stop"
    post:
      operationId: createStop
      tags: [stops]
      summary: Create a stop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopInput"
      responses:
        "201":
          description: Created stop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stop"
  /stops/{id}:
    get:
      operationId: getStop
      tags: [stops]
      summary: Get a stop by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stop"
    patch:
      operationId: updateStop
      tags: [stops]
      summary: Update a stop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StopUpdate"
      responses:
        "200":
          description: Updated stop
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stop"
    delete:
      operationId: deleteStop
      tags: [stops]
      summary: Delete a stop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted

  # --- TIMETABLES ---
...
  /reports/daily:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Daily passenger totals
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /reports/weekly:
    get:
      operationId: getWeeklyReport
      tags: [reports]
      summary: Weekly passenger totals
      parameters:
        - name: weekStart
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Weekly report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyReport"
  /reports/monthly:
    get:
      operationId: getMonthlyReport
      tags: [reports]
      summary: Monthly passenger totals
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: month
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
  /reports/by-duty:
    get:
      operationId: getReportByDuty
      tags: [reports]
      summary: Passenger totals grouped by duty
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Duty breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DutyBreakdown"
  /reports/by-stop:
    get:
      operationId: getReportByStop
      tags: [reports]
      summary: Passenger totals grouped by stop
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Stop breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StopBreakdown"
  /reports/by-hour:
    get:
      operationId: getReportByHour
      tags: [reports]
      summary: Passenger counts grouped by hour
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Hourly breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HourlyBreakdown"
  /reports/missing-submissions:
    get:
      operationId: getMissingSubmissions
      tags: [reports]
      summary: List departures with missing passenger count submissions
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Missing submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MissingSubmission"
  /reports/late-submissions:
    get:
      operationId: getLateSubmissions
      tags: [reports]
      summary: List late passenger count submissions
      parameters:
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Late submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LateSubmission"
  /reports/export:
    get:
      operationId: exportReport
      tags: [reports]
      summary: Export passenger records as CSV
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [csv]
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema:
                type: string

  # --- USERS ---
...
components:
  schemas:
...
    Stop:
      type: object
      required: [id, name, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        geofenceRadiusMeters:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
    StopInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        geofenceRadiusMeters:
          type: integer
    StopUpdate:
      type: object
      properties:
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        geofenceRadiusMeters:
          type: integer
        isActive:
          type: boolean

    # --- TIMETABLE ---
...
    DailyReport:
      type: object
      required: [date, totalPassengers, totalDepartures, completedDepartures, missingDepartures, lateDepartures, byDuty, byStop, byHour]
      properties:
        date:
          type: string
        totalPassengers:
          type: integer
        totalDepartures:
          type: integer
        completedDepartures:
          type: integer
        missingDepartures:
          type: integer
        lateDepartures:
          type: integer
        busiestDeparture:
          type: ["string", "null"]
        quietestDeparture:
          type: ["string", "null"]
        averagePassengers:
          type: number
        byDuty:
          type: array
          items:
            $ref: "#/components/schemas/DutyBreakdown"
        byStop:
          type: array
          items:
            $ref: "#/components/schemas/StopBreakdown"
        byHour:
          type: array
          items:
            $ref: "#/components/schemas/HourlyBreakdown"
    WeeklyReport:
      type: object
      required: [weekStart, weekEnd, totalPassengers, days]
      properties:
        weekStart:
          type: string
        weekEnd:
          type: string
        totalPassengers:
          type: integer
        averageDaily:
          type: number
        days:
          type: array
          items:
            type: object
            required: [date, totalPassengers]
            properties:
              date:
                type: string
              totalPassengers:
                type: integer
    MonthlyReport:
      type: object
      required: [year, month, totalPassengers, weeks]
      properties:
        year:
          type: integer
        month:
          type: integer
        totalPassengers:
          type: integer
        averageDaily:
          type: number
        weeks:
          type: array
          items:
            type: object
            required: [weekStart, totalPassengers]
            properties:
              weekStart:
                type: string
              totalPassengers:
                type: integer
    DutyBreakdown:
      type: object
      required: [dutyId, dutyName, totalPassengers, completedDepartures, totalDepartures]
      properties:
        dutyId:
          type: integer
        dutyName:
          type: string
        totalPassengers:
          type: integer
        completedDepartures:
          type: integer
        totalDepartures:
          type: integer
        averagePassengers:
          type: number
    StopBreakdown:
      type: object
      required: [stopId, stopName, totalPassengers, totalDepartures]
      properties:
        stopId:
          type: integer
        stopName:
          type: string
        totalPassengers:
          type: integer
        totalDepartures:
          type: integer
        averagePassengers:
          type: number
    HourlyBreakdown:
      type: object
      required: [hour, totalPassengers, departures]
      properties:
        hour:
          type: integer
        totalPassengers:
          type: integer
        departures:
          type: integer
    MissingSubmission:
      type: object
      required: [timetableEntryId, dutyId, dutyName, stopId, stopName, scheduledTime, date]
      properties:
        timetableEntryId:
          type: integer
        dutyId:
          type: integer
        dutyName:
          type: string
        stopId:
          type: integer
        stopName:
          type: string
        scheduledTime:
          type: string
        date:
          type: string
        driverId:
          type: ["integer", "null"]
        driverName:
          type: ["string", "null"]
    LateSubmission:
      type: object
      required: [id, dutyId, dutyName, stopId, stopName, scheduledTime, actualSubmissionTime, minutesLate, passengerCount, date]
      properties:
        id:
          type: integer
        dutyId:
          type: integer
        dutyName:
          type: string
        stopId:
          type: integer
        stopName:
          type: string
        scheduledTime:
          type: string
        actualSubmissionTime:
          type: string
        minutesLate:
          type: integer
        passengerCount:
          type: integer
        date:
          type: string
        driverName:
          type: ["string", "null"]

    # --- USERS ---
...