# lib/api-zod/src/generated/types/stopUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Swans ShuttleHub API
 * OpenAPI spec version: 0.1.0
 */

export interface StopUpdate {
  name?: string;
  latitude?: number;
  longitude?: number;
  geofenceRadiusMeters?: number;
  isActive?: boolean;
}